Refactor tokenizers
